+++ packages/vite-plugin/test/no-macro-error/entrypoint.js
@@ -1,0 +1,5 @@
+import { t } from "@lingui/macro"
+
+export async function load() {
+  return t`Ola`
+}